Added Coinbase connection support
